requirement.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-core==2.5.0
google-cloud-firestore==2.24.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.5
rsa==4.9.1
typing_extensions==4.15.0
urllib3==2.6.3
